requirements.txt
langchain==0.3.26
langchain_community==0.3.27
neo4j==5.28.1
pydantic==2.11.7
python-dotenv==1.1.1
streamlit==1.46.1
